requirements/base.txt
fastapi
uvicorn
sqlalchemy
asyncpg
passlib[bcrypt]
pyjwt[crypto]
